# optimize internal parsing function -- avoid unsafe []byte parsing