--- a/marshal.go
+++ b/marshal.go
@@ -4,10 +4,7 @@
 
 package uuid
 
-import (
-	"fmt"
-	"unsafe"
-)
+import "fmt"
 
 // MarshalText implements encoding.TextMarshaler.
 func (u UUID) MarshalText() ([]byte, error) {
@@ -20,7 +17,7 @@ func (u UUID) MarshalText() ([]byte, error) {
 func (u *UUID) UnmarshalText(data []byte) error {
 	// See comment in ParseBytes why we do this.
 	// id, err := ParseBytes(data)
-        id, err := Parse(*(*string)(unsafe.Pointer(&data)))
+	id, err := ParseBytes(data)
 	if err == nil {
 		*u = id
 	}